refactor(FaceManagmentRouter): Extract all the face managment entrypoints to independed router
